+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 Nothing yet.
 
+## 1.1.0 - Thu 28th Jan 2016
+
+### New
+
+* Added the basic exception hierarchy
+  - Added `Exxx_ArrayToolsException`
+  - Added `E4xx_ArrayToolsException`
+  - Added `E4xx_UnsupportedType`
+* Added support for converting key/value pairs in a string into an array
+  - Added `Parsers\ConvertKeyValuePairsToArray`
+
 ## 1.0.0 - Thu 28th Jan 2016
 
 ### New